+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/tests/test_custom.py
@@ -13,8 +13,8 @@ import yaml
 
 from msrestazure.azure_exceptions import CloudError
 from azure.graphrbac.models import GraphErrorException
-from azure.cli.command_modules.acs._params import (regionsInPreview,
-                                                   regionsInProd)
+from azure.cli.command_modules.acs._params import (regions_in_preview,
+                                                   regions_in_prod)
 from azure.cli.command_modules.acs.custom import (merge_kubernetes_configurations, list_acs_locations,
                                                   _acs_browse_internal, _add_role_assignment, _get_default_dns_prefix,
                                                   create_application)
@@ -26,11 +26,12 @@ from azure.cli.core.util import CLIError
 
 class AcsCustomCommandTest(unittest.TestCase):
     def test_list_acs_locations(self):
-        regions = list_acs_locations()
+        client, cmd = mock.MagicMock(), mock.MagicMock()
+        regions = list_acs_locations(client, cmd)
         prodregions = regions["productionRegions"]
         previewregions = regions["previewRegions"]
-        self.assertListEqual(prodregions, regionsInProd, "Production regions doesn't match")
-        self.assertListEqual(previewregions, regionsInPreview, "Preview regions doesn't match")
+        self.assertListEqual(prodregions, regions_in_prod, "Production regions doesn't match")
+        self.assertListEqual(previewregions, regions_in_preview, "Preview regions doesn't match")
 
     def test_get_default_dns_prefix(self):
         name = 'test5678910'
@@ -47,16 +48,18 @@ class AcsCustomCommandTest(unittest.TestCase):
     def test_add_role_assignment_basic(self):
         role = 'Owner'
         sp = '1234567'
+        cli_ctx = mock.MagicMock()
 
         with mock.patch(
                 'azure.cli.command_modules.acs.custom.create_role_assignment') as create_role_assignment:
-            ok = _add_role_assignment(role, sp, delay=0)
-            create_role_assignment.assert_called_with(role, sp)
+            ok = _add_role_assignment(cli_ctx, role, sp, delay=0)
+            create_role_assignment.assert_called_with(cli_ctx, role, sp)
             self.assertTrue(ok, 'Expected _add_role_assignment to succeed')
 
     def test_add_role_assignment_exists(self):
         role = 'Owner'
         sp = '1234567'
+        cli_ctx = mock.MagicMock()
 
         with mock.patch(
                 'azure.cli.command_modules.acs.custom.create_role_assignment') as create_role_assignment:
@@ -66,14 +69,15 @@ class AcsCustomCommandTest(unittest.TestCase):
             err = CloudError(resp)
             err.message = 'The role assignment already exists.'
             create_role_assignment.side_effect = err
-            ok = _add_role_assignment(role, sp, delay=0)
+            ok = _add_role_assignment(cli_ctx, role, sp, delay=0)
 
-            create_role_assignment.assert_called_with(role, sp)
+            create_role_assignment.assert_called_with(cli_ctx, role, sp)
             self.assertTrue(ok, 'Expected _add_role_assignment to succeed')
 
     def test_add_role_assignment_fails(self):
         role = 'Owner'
         sp = '1234567'
+        cli_ctx = mock.MagicMock()
 
         with mock.patch(
                 'azure.cli.command_modules.acs.custom.create_role_assignment') as create_role_assignment:
@@ -83,9 +87,9 @@ class AcsCustomCommandTest(unittest.TestCase):
             err = CloudError(resp)
             err.message = 'Internal Error'
             create_role_assignment.side_effect = err
-            ok = _add_role_assignment(role, sp, delay=0)
+            ok = _add_role_assignment(cli_ctx, role, sp, delay=0)
 
-            create_role_assignment.assert_called_with(role, sp)
+            create_role_assignment.assert_called_with(cli_ctx, role, sp)
             self.assertFalse(ok, 'Expected _add_role_assignment to fail')
 
     @mock.patch('azure.cli.command_modules.acs.custom._get_subscription_id')
@@ -93,13 +97,14 @@ class AcsCustomCommandTest(unittest.TestCase):
         acs_info = ContainerService("location", {}, {}, {})
         acs_info.orchestrator_profile = ContainerServiceOrchestratorProfile(
             ContainerServiceOrchestratorTypes.kubernetes)
+        client, cmd = mock.MagicMock(), mock.MagicMock()
 
         with mock.patch('azure.cli.command_modules.acs.custom._get_acs_info',
                         return_value=acs_info) as get_acs_info:
             with mock.patch(
                     'azure.cli.command_modules.acs.custom._k8s_browse_internal') as k8s_browse:
-                _acs_browse_internal(acs_info, 'resource-group', 'name', False, 'ssh/key/file')
-                get_acs_info.assert_called_with('name', 'resource-group')
+                _acs_browse_internal(client, cmd, acs_info, 'resource-group', 'name', False, 'ssh/key/file')
+                get_acs_info.assert_called_once()
                 k8s_browse.assert_called_with('name', acs_info, False, 'ssh/key/file')
 
     @mock.patch('azure.cli.command_modules.acs.custom._get_subscription_id')
@@ -107,10 +112,11 @@ class AcsCustomCommandTest(unittest.TestCase):
         acs_info = ContainerService("location", {}, {}, {})
         acs_info.orchestrator_profile = ContainerServiceOrchestratorProfile(
             ContainerServiceOrchestratorTypes.dcos)
+        client, cmd = mock.MagicMock(), mock.MagicMock()
 
         with mock.patch(
                 'azure.cli.command_modules.acs.custom._dcos_browse_internal') as dcos_browse:
-            _acs_browse_internal(acs_info, 'resource-group', 'name', False, 'ssh/key/file')
+            _acs_browse_internal(client, cmd, acs_info, 'resource-group', 'name', False, 'ssh/key/file')
             dcos_browse.assert_called_with(acs_info, False, 'ssh/key/file')
 
     def test_merge_credentials_non_existent(self):